frontend: split out the search map presentation functionality
Creates a SMMRealtime to share with other map realtime components
when they are converted.
